% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aad.R
\name{aad}
\alias{aad}
\title{Mean absolute deviation}
\usage{
aad(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector containing the observations.}

\item{na.rm}{Set to \code{TRUE} to remove missing values before computing
the deviation. Defaults to \code{FALSE}.}
}
\value{
A single number giving the mean absolute deviation.
}
\description{
Calculates the mean absolute deviation from the sample mean.
}
\details{
Computes the average of the absolute differences between each
observation and the sample mean of \code{x}, i.e.
\code{mean(abs(x - mean(x)))}.
}
\examples{
x <- c(1, 3, 6)
aad(x)

# missing values
x <- c(1, 3, NA, 6)
aad(x) # returns NA
aad(x, na.rm = TRUE) # ignores the missing value

}
\seealso{
\code{\link{mean}}, \code{\link{sd}}, \code{\link{var}}
}
