% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permuteLevels.R
\name{permuteLevels}
\alias{permuteLevels}
\title{Permute the levels of a factor}
\usage{
permuteLevels(x, perm, ordered = is.ordered(x), invert = FALSE)
}
\arguments{
\item{x}{A factor.}

\item{perm}{An integer vector of the same length as \code{nlevels(x)}.
When \code{invert = FALSE} (the default), \code{perm[k]} gives the
index of the old level that should appear in position \code{k} of
the new ordering. When \code{invert = TRUE}, \code{perm[k]} gives
the new position to assign to the \code{k}-th old level.}

\item{ordered}{Set to \code{TRUE} to return an ordered factor. Defaults
to \code{is.ordered(x)}, preserving the ordered status of the input.}

\item{invert}{Set to \code{TRUE} to apply the inverse of \code{perm}.
See the \code{perm} description above and the examples below.}
}
\value{
A factor with the same values as \code{x} but with the levels
in the new order specified by \code{perm}.
}
\description{
Reorder the levels of a factor into any order you specify.
}
\details{
Similar to \code{\link{relevel}}, but more general:
\code{relevel} can only move one level to the front, whereas
\code{permuteLevels} can place the levels in any order. This is useful
when you want to control the order in which levels appear on a plot axis
or in a table.
}
\examples{
# factor with levels a, b, c, d, e, f (in that order)
x <- factor(c(1, 4, 2, 2, 3, 3, 5, 5, 6, 6), labels = letters[1:6])
levels(x)

# move level e to position 1, c to position 2, b to 3, a to 4, d to 5, f to 6
permuteLevels(x, perm = c(5, 3, 2, 1, 4, 6))

# using invert = TRUE: move level a to position 5, b to 3, c to 2, etc.
permuteLevels(x, perm = c(5, 3, 2, 1, 4, 6), invert = TRUE)

}
\seealso{
\code{\link{factor}},
\code{\link{relevel}},
\code{\link{order}}
}
