% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramersV.R
\name{cramersV}
\alias{cramersV}
\title{Cramer's V}
\usage{
cramersV(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{chisq.test}}, in the same format
accepted by that function.}
}
\value{
A single number giving the value of Cramer's V.
}
\description{
Calculates Cramer's V, a measure of the strength of association
for chi-square tests.
}
\details{
Cramer's V summarises the strength of association from a chi-square
test. It is appropriate for both tests of association (two categorical
variables) and goodness of fit tests (one variable versus hypothesised
probabilities). Values range from 0 (no association) to 1 (perfect
association). The arguments are passed directly to \code{\link{chisq.test}},
so the input format is the same.
}
\examples{
# frequency table for two groups, each choosing from three options
condition1 <- c(30, 20, 50)
condition2 <- c(35, 30, 35)
X <- cbind(condition1, condition2)
rownames(X) <- c("choice1", "choice2", "choice3")

# chi-square test of association
chisq.test(X)

# effect size estimate
cramersV(X)

}
\seealso{
\code{\link{chisq.test}},
\code{\link{associationTest}},
\code{\link{goodnessOfFitTest}}
}
