\name{lsr-package}
\alias{lsr-package}
\alias{lsr}
\docType{package}

\title{
Companion to "Learning Statistics with R"
}

\description{
A collection of tools intended to make introductory statistics easier to teach, including wrappers for common hypothesis tests and basic data manipulation. It accompanies Navarro, D. J. (2015). Learning Statistics with R: A Tutorial for Psychology Students and Other Beginners, Version 0.6.
}

\details{
\tabular{ll}{
Package: \tab lsr\cr
Type: \tab Package\cr
Version: \tab 0.5.1\cr
License: \tab MIT \cr
LazyLoad: \tab yes\cr
}

The lsr package provide a few tools that are useful for students new both to statistics and to R. For example, it includes separate wrapper functions for independent samples t-tests and paired samples t-test, each of which produce output that state the null and alternative hypotheses in a verbose fashion. Other tools are provided to assist students with data manipulation.
}

\author{
Author: Danielle Navarro
\cr
Maintainer: Daniellr Navarro <d.navarro@unsw.edu.au>
}

\references{
Navarro, D. J. (2015). \emph{Learning Statistics with R: A Tutorial for Psychology Students and Other Beginners}, Version 0.6. University of New South Wales. Sydney, Australia. URL: \url{https://learningstatisticswithr.com}
}

\keyword{ package }

%\seealso{
%}

%\examples{
%## ~~ simple examples of the most important functions ~~
%}
