\name{lsr-package}
\alias{lsr-package}
\alias{lsr}
\docType{package}

\title{
Companion to "Learning Statistics with R"
}

\description{
A collection of tools intended to make introductory statistics easier to teach, including wrappers for common hypothesis tests and basic data manipulation. It accompanies Navarro, D. J. (2015). \emph{Learning Statistics with R: A Tutorial for Psychology Students and Other Beginners}, Version 0.5. [Lecture notes] School of Psychology, University of Adelaide, Adelaide, Australia. ISBN: 978-1-326-18972-3. URL: \url{http://health.adelaide.edu.au/psychology/ccs/teaching/lsr/}.
}

\details{
\tabular{ll}{
Package: \tab lsr\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2015-03-01\cr
License: \tab GPL-3 \cr
LazyLoad: \tab yes\cr
}

The lsr package provide a few tools that are useful for students new both to statistics and to R. For example, it includes separate wrapper functions for independent samples t-tests and paired samples t-test, each of which produce output that state the null and alternative hypotheses in a verbose fashion. Other tools are provided to assist students with data manipulation. 

WARNING: This package is still under development. The functions work for most situations that I envisage them being used for, but some extensions are planned. In particular, for the moment there are no guarantees of backwards compatibility. Until this notice disappears some care is advised.

}

\author{
Author: Daniel Navarro
\cr
Maintainer: Daniel Navarro <daniel.navarro@adelaide.edu.au>
}

\references{
Navarro, D. J. (2015). \emph{Learning Statistics with R: A Tutorial for Psychology Students and Other Beginners}, Version 0.5. [Lecture notes] School of Psychology, University of Adelaide, Adelaide, Australia. ISBN: 978-1-326-18972-3. URL: \url{http://health.adelaide.edu.au/psychology/ccs/teaching/lsr/.}
}

\keyword{ package }

%\seealso{
%}

%\examples{
%## ~~ simple examples of the most important functions ~~
%}
