% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associationTest.R
\name{print.assocTest}
\alias{print.assocTest}
\title{Print method for lsr chi-square tests}
\usage{
\method{print}{assocTest}(x, ...)
}
\arguments{
\item{x}{An object of class 'assocTest'}

\item{...}{For consistency with the generic (unused)}
}
\value{
Invisibly returns the original object
}
\description{
Print method for lsr chi-square tests
}
