% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{print.correlate}
\alias{print.correlate}
\title{Print method for correlate objects}
\usage{
\method{print}{correlate}(x, ...)
}
\arguments{
\item{x}{An object of class 'correlate'}

\item{...}{For consistency with the generic (unused)}
}
\value{
Invisibly returns the original object
}
\description{
Print method for correlate objects
}
