% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessOfFitTest.R
\name{print.gofTest}
\alias{print.gofTest}
\title{Print method for lsr goodness-of-fit tests}
\usage{
\method{print}{gofTest}(x, ...)
}
\arguments{
\item{x}{An object of class 'gofTest'}

\item{...}{For consistency with the generic (unused)}
}
\value{
Invisibly returns the original object
}
\description{
Print method for lsr goodness-of-fit tests
}
