% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTTest.R
\name{print.TTest}
\alias{print.TTest}
\title{Print method for lsr t-tests}
\usage{
\method{print}{TTest}(x, ...)
}
\arguments{
\item{x}{An object of class 'TTest'}

\item{...}{For consistency with the generic (unused)}
}
\value{
Invisibly returns the original object
}
\description{
Print method for lsr t-tests
}
