% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{print.whoList}
\alias{print.whoList}
\title{Print method for whoList objects}
\usage{
\method{print}{whoList}(x, ...)
}
\arguments{
\item{x}{An object of class 'whoList'}

\item{...}{For consistency with the generic (unused)}
}
\value{
Invisibly returns the original object
}
\description{
Print method for whoList objects
}
